packages/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
geoalchemy2==0.15.2
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
redis==5.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
prometheus-fastapi-instrumentator==7.0.0
python-json-logger==3.2.1
